schemas.pyをschemas配下のファイルに分割
