+++ app/schemas/booking.py
@@ -1,0 +1,15 @@
+import datetime
+from pydantic import BaseModel
+
+class BookingCreate(BaseModel):
+    user_id: int
+    room_id: int
+    booked_num: int
+    start_datetime: datetime.datetime
+    end_datetime: datetime.datetime
+
+class Booking(BookingCreate):
+    booking_id: int
+
+    class Config:
+        orm_mode = True
